requirements.txt
Flask==3.1.0
gunicorn==23.0.0
google-cloud-speech==2.31.0
google-generativeai==0.8.4
google-cloud-texttospeech==2.25.1
PyMuPDF==1.25.5
google-cloud-storage==2.14.0
